refactor nwm request class into module
